pyproject.toml
[project]
name = "ley-ar"
version = "0.1.0"
description = "MCP server de legislacion laboral argentina para agentes de IA"
readme = "readme.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Luciano Sanchez", email = "luchoqq@gmail.com" },
]
keywords = ["mcp", "legal", "argentina", "laboral", "legislacion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp>=1.0.0",
    "python-dateutil>=2.8.0",
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/LuchoQQ/ley-ar_mcp"
Repository = "https://github.com/LuchoQQ/ley-ar_mcp"

[project.scripts]
ley-ar = "ley_ar.server:mcp.run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ley_ar"]

[tool.hatch.build.targets.wheel.force-include]
"src/ley_ar/data/legislacion" = "ley_ar/data/legislacion"
"src/ley_ar/data/descriptores" = "ley_ar/data/descriptores"

[tool.hatch.build.targets.sdist]
include = ["src/ley_ar"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
